+++ tests/fkooman/Http/UrlTest.php
@@ -25,11 +25,10 @@ class UrlTest extends PHPUnit_Framework_TestCase
     public function testHttp()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '80',
-            'QUERY_STRING' => 'foo=bar',
             'REQUEST_URI' => '/bar/index.php?foo=bar',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -46,12 +45,10 @@ class UrlTest extends PHPUnit_Framework_TestCase
     public function testHttps()
     {
         $srv = array(
-            'HTTPS' => 'on',
+            'REQUEST_SCHEME' => 'https',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '443',
-            'QUERY_STRING' => 'foo=bar',
             'REQUEST_URI' => '/bar/index.php?foo=bar',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -67,16 +64,13 @@ class UrlTest extends PHPUnit_Framework_TestCase
         $this->assertSame('https://www.example.org/bar/index.php?foo=bar', $u->toString());
     }
 
-    public function testHttpsOpenVpnPortSharing()
+    public function testHttpsNonStandardPort()
     {
         $srv = array(
-            'HTTPS' => 'on',
-            'HTTP_HOST' => 'www.example.org',
+            'REQUEST_SCHEME' => 'https',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '8443',
-            'QUERY_STRING' => '',
             'REQUEST_URI' => '/bar/index.php',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -85,17 +79,16 @@ class UrlTest extends PHPUnit_Framework_TestCase
         $this->assertSame(8443, $u->getPort());
         $this->assertSame('/', $u->getPathInfo());
         $this->assertSame('/bar/', $u->getRoot());
-        $this->assertSame('https://www.example.org/bar/', $u->getRootUrl());
+        $this->assertSame('https://www.example.org:8443/bar/', $u->getRootUrl());
     }
 
     public function testHttpNonStandardPort()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '8080',
-            'QUERY_STRING' => 'foo=bar',
             'REQUEST_URI' => '/bar/index.php?foo=bar',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -112,12 +105,11 @@ class UrlTest extends PHPUnit_Framework_TestCase
     public function testHttpPathInfo()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '80',
-            'QUERY_STRING' => 'foo=bar',
             'PATH_INFO' => '/def',
             'REQUEST_URI' => '/bar/index.php/def?foo=bar',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -134,12 +126,11 @@ class UrlTest extends PHPUnit_Framework_TestCase
     public function testHttpServerRewrite()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '80',
-            'QUERY_STRING' => 'foo=bar',
             'PATH_INFO' => '/def',
             'REQUEST_URI' => '/bar/def?foo=bar',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -156,12 +147,11 @@ class UrlTest extends PHPUnit_Framework_TestCase
     public function testHttpServerRewriteRoot()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '80',
-            'QUERY_STRING' => 'foo=bar',
             'PATH_INFO' => '/def',
             'REQUEST_URI' => '/def?foo=bar',
-            'SCRIPT_NAME' => '/index.php',
         );
 
         $u = new Url($srv);
@@ -178,11 +168,10 @@ class UrlTest extends PHPUnit_Framework_TestCase
     public function testEmptyQueryString()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '80',
-            'QUERY_STRING' => '',
             'REQUEST_URI' => '/bar/index.php',
-            'SCRIPT_NAME' => '/bar/index.php',
         );
 
         $u = new Url($srv);
@@ -196,44 +185,23 @@ class UrlTest extends PHPUnit_Framework_TestCase
         $this->assertNull($u->getQueryParameter('foo'));
     }
 
-    public function testHttpsProxy()
-    {
-        $srv = array(
-            'SERVER_NAME' => 'www.example.org',
-            'SERVER_PORT' => '80',
-            'HTTP_X_FORWARDED_PROTO' => 'https',
-            'QUERY_STRING' => '',
-            'REQUEST_URI' => '/bar/index.php',
-            'SCRIPT_NAME' => '/bar/index.php',
-        );
-
-        $u = new Url($srv);
-        $this->assertSame('https', $u->getScheme());
-        $this->assertSame('www.example.org', $u->getHost());
-        $this->assertSame(80, $u->getPort());
-        $this->assertSame('/bar/', $u->getRoot());
-        $this->assertSame('/', $u->getPathInfo());
-        $this->assertSame('https://www.example.org/bar/', $u->getRootUrl());
-    }
-
     /**
      * @expectedException RuntimeException
-     * @expectedExceptionMessage missing key "SERVER_NAME"
+     * @expectedExceptionMessage missing key "REQUEST_SCHEME"
      */
     public function testMissingKey()
     {
-        $u = new Url(array());
+        $u = new Url([]);
     }
 
     public function testScriptNameFix()
     {
         $srv = array(
+            'REQUEST_SCHEME' => 'http',
             'SERVER_NAME' => 'www.example.org',
             'SERVER_PORT' => '80',
-            'QUERY_STRING' => 'foo=bar',
             'PATH_INFO' => '/foo',
             'REQUEST_URI' => '/bar/index.php/foo?foo=bar',
-            'SCRIPT_NAME' => '/bar/index.php/foo',  // mistakenly includes PATH_INFO
         );
 
         $u = new Url($srv);
@@ -245,23 +213,5 @@ class UrlTest extends PHPUnit_Framework_TestCase
         $this->assertSame('http://www.example.org/bar/index.php/', $u->getRootUrl());
         $this->assertSame(array('foo' => 'bar'), $u->getQueryStringAsArray());
         $this->assertSame('bar', $u->getQueryParameter('foo'));
-    }
-
-    /**
-     * @expectedException InvalidArgumentException
-     * @expectedExceptionMessage non ASCII characters detected
-     */
-    public function testNonAsciiUrlPart()
-    {
-        $srv = array(
-            'SERVER_NAME' => 'www.example.org',
-            'SERVER_PORT' => '80',
-            'QUERY_STRING' => 'name=François',
-            'PATH_INFO' => '/foo',
-            'REQUEST_URI' => '/bar/index.php/foo?name=François',
-            'SCRIPT_NAME' => '/bar/index.php',
-        );
-
-        $u = new Url($srv);
     }
 }